+++ ricaun.NUnit.Tests/NUnitTests.cs
@@ -19,7 +19,7 @@ namespace ricaun.NUnit.Tests
         }
 
         [Test]
-        public void TestAssemblyInitialize()
+        public void TestAssembly_Initialize()
         {
             TestEngine.Initialize(out string message);
             Console.WriteLine(message);
@@ -29,14 +29,14 @@ namespace ricaun.NUnit.Tests
         }
 
         [Test]
-        public void TestAssemblyContainNUnit()
+        public void TestAssembly_ContainNUnit()
         {
             Console.WriteLine(fileName);
             Assert.IsTrue(TestEngine.ContainNUnit(pathFile));
         }
 
         [Test]
-        public void TestAssemblyNames()
+        public void TestAssembly_Names()
         {
             Console.WriteLine(fileName);
             var names = TestEngine.GetTestFullNames(pathFile);
@@ -47,7 +47,7 @@ namespace ricaun.NUnit.Tests
         }
 
         [Test]
-        public void TestAssemblyNamesAlias()
+        public void TestAssembly_NamesAlias()
         {
             Console.WriteLine(fileName);
             var testModel = TestEngine.TestAssembly(pathFile);
@@ -67,7 +67,7 @@ namespace ricaun.NUnit.Tests
         }
 
         [Test]
-        public void TestAssemblyNamesAlias_FullName()
+        public void TestAssembly_NamesAlias_FullName()
         {
             Console.WriteLine(fileName);
             var testModel = TestEngine.TestAssembly(pathFile);
@@ -77,11 +77,30 @@ namespace ricaun.NUnit.Tests
             {
                 Console.WriteLine(testFullName);
             }
-            Assert.IsTrue(nameAlias.SequenceEqual(testFullNames));
+            Assert.IsTrue(nameAlias.SequenceEqual(testFullNames), "Sequence Alias and FullName equal");
         }
 
         [Test]
-        public void TestAssemblyService()
+        public void TestAssembly_TestEngineResult()
+        {
+            var testCount = 0;
+            var textOut = "";
+            Console.WriteLine(fileName);
+            TestEngineResult.Result = new TestModelResult((test) =>
+            {
+                testCount++;
+                textOut += $"{testCount}\t{test.FullName}\n";
+            });
+            var testModel = TestEngine.TestAssembly(pathFile);
+            //Console.WriteLine(testModel.AsString());
+            Console.WriteLine(textOut);
+            TestEngineResult.Result = null;
+
+            Assert.AreEqual(testModel.TestCount, testCount);
+        }
+
+        [Test]
+        public void TestAssembly_Service()
         {
             Console.WriteLine(fileName);
             var service = new Services.TestAssemblyService(pathFile);
@@ -105,7 +124,7 @@ namespace ricaun.NUnit.Tests
         }
 
         [Test]
-        public void TestAssemblyExplicit()
+        public void TestAssembly_Explicit()
         {
             Console.WriteLine(fileName);
             TestEngineFilter.ExplicitEnabled = true;
@@ -118,7 +137,7 @@ namespace ricaun.NUnit.Tests
         }
 
         [Test]
-        public void TestAssemblyParameters()
+        public void TestAssembly_Parameters()
         {
             Console.WriteLine(fileName);
             TestEngineFilter.Add("*.TestParameter*");
@@ -137,7 +156,7 @@ namespace ricaun.NUnit.Tests
         [TestCase("*.TestCases(?)", 2)]
         [TestCase("*.TestSame?", 2)]
         [TestCase("*", 32)]
-        public void TestAssemblyFilter(string testName, int numberOfTests)
+        public void TestAssembly_Filter(string testName, int numberOfTests)
         {
             TestEngineFilter.Add(testName);
             var testModel = TestEngine.TestAssembly(pathFile);
@@ -151,7 +170,7 @@ namespace ricaun.NUnit.Tests
         [TestCase(1)]
         [TestCase(2)]
         [TestCase(4)]
-        public void TestAssemblyFilterName(int numberOfTests)
+        public void TestAssembly_FilterName(int numberOfTests)
         {
             Console.WriteLine(fileName);
             var names = TestEngine.GetTestFullNames(pathFile);